### 2024-07-19 #7014 @editorialbot
[PRE REVIEW]: ngsPETSc: A coupling between NETGEN/NGSolve and PETSc
**Submitting author:** @uzerbinati (<a href="http://orcid.org/0000-0002-2577-1106">Umberto Zerbinati</a>)
**Repository:** https://github.com/NGSolve/ngsPETSc
**Branch with paper.md** (empty if default branch):
**Version:** v.0.0.5
