### 2024-07-19 #7014 @danielskatz
👋 @UZerbinati - your readme appears blank, which will need to be changed to satisfy JOSS review criteria, such as in https://joss.readthedocs.io/en/latest/review_criteria.html#documentation
