### 2024-07-19 #7014 @UZerbinati
Hi @danielskatz ! Thank you for letting me know: I've updated the README to include a statement of need, installation instructions and a minimal example on how to test the code.
Also, I now realise that I miss-typed the title of the paper during submission, the correct title should be: "
