### 2024-07-21 39b5ba8
Updated DOIs
Signed-off-by: Umberto Zerbinati <zerbinati@maths.ox.ac.uk>
