+++ paper/paper.bib
@@ -1,12 +1,13 @@
 @article{Netgen,
-pages = {41-52},
-publisher = {Springer Science and Business Media LLC},
-volume = {1},
-year = {1997},
-title = {NETGEN An advancing front 2D/3D-mesh generator based on abstract rules},
-author = {Schöberl, Joachim},
-journal = {Computing and visualization in science},
-number = {1},
+  pages = {41-52},
+  publisher = {Springer Science and Business Media LLC},
+  volume = {1},
+  year = {1997},
+  title = {NETGEN An advancing front 2D/3D-mesh generator based on abstract rules},
+  author = {Schöberl, Joachim},
+  journal = {Computing and visualization in science},
+  number = {1},
+  doi = {10.1007/s007910050004}
 }
 @article{NGSolve,
   title={C++ 11 implementation of finite elements in NGSolve},
@@ -19,8 +20,9 @@ number = {1},
   title={Firedrake user manual},
   author={Ham, David A and Kelly, Paul HJ and Mitchell, Lawrence and Cotter, Colin J and Kirby, Robert C and Sagiyama, Koki and Bouziani, Nacime and Vorderwuelbecke, Sophia and Gregory, Thomas J and Betteridge, Jack and others},
   journal={Imperial College London and University of Oxford and Baylor University and University of Washington,},
-  volume={5},
-  year={2023}
+  month={5},
+  year={2023},
+  doi= {10.25561/104839}
 }
 @techreport{PETSc,
   author={Satish Balay and Shrirang Abhyankar and Mark~F. Adams and Steven Benson and Jed
@@ -35,7 +37,8 @@ number = {1},
   title={{PETSc/TAO} Users Manual},
   institution={Argonne National Laboratory},
   number={ANL-21/39 - Revision 3.20},
-  year={2023}
+  year={2023},
+  doi={10.2172/2205494}
 }
 @article{petsc4py,
   title={Parallel distributed computing using Python},
@@ -45,7 +48,8 @@ number = {1},
   number={9},
   pages={1124 - 1139},
   issn={0309-1708},
-  year={2011}
+  year={2011},
+  doi={10.1016/j.advwatres.2011.04.013}
 }
 @misc{manual,
   author       = {Zerbinati, Umberto},
@@ -55,7 +59,6 @@ number = {1},
   publisher    = {Zenodo},
   version      = {0.0.5},
   doi          = {10.5281/zenodo.12650574},
-  url          = {https://ngspetsc.readthedocs.io/en/latest/}
 }
 @article{PETScBDDC,
       author = {Stefano Zampini},
@@ -66,8 +69,6 @@ number = {1},
       pages = {S282-S306},
       year = {2016},
       doi = {10.1137/15M1025785},
-      URL = {http://dx.doi.org/10.1137/15M1025785},
-      eprint = {http://dx.doi.org/10.1137/15M1025785}
 }
 @article{PETScGAMG,
 title = {Parallel multigrid smoothing: polynomial versus {Gauss--Seidel}},
@@ -77,9 +78,7 @@ number = {2},
 pages = {593-610},
 year = {2003},
 issn = {0021-9991},
-doi = {https://doi.org/10.1016/S0021-9991(03)00194-3},
-url = {https://www.sciencedirect.com/science/article/pii/S0021999103001943},
-author = {Mark Adams and Marian Brezina and Jonathan Hu and Ray Tuminaro},
+doi = {10.1016/S0021-9991(03)00194-3},
 }
 @article{BenziOlshanskii,
 author = { Benzi, Michele and  Olshanskii, Maxim A.},
@@ -89,6 +88,7 @@ volume = {28},
 number = {6},
[…]
