### 2024-07-21 #7014 @UZerbinati
I've added DOIs for the papers I could find a DOI for.
I suspect there are no DOIs for the remaining papers.
