### 2024-07-31 42f33cc
Fixed missing return identified by Paul Schwering.
Signed-off-by: Umberto Zerbinati <zerbinati@maths.ox.ac.uk>
