### 2024-09-06 5b7fe51
Added labels atribute to mesh, makes it easier :)
Signed-off-by: Umberto Zerbinati <zerbinati@maths.ox.ac.uk>
