+++ ngsPETSc/plex.py
@@ -199,16 +199,17 @@ class MeshMapping:
                 plex.setName(self.name)
                 vStart, _ = plex.getDepthStratum(0)   # vertices
                 nMat = len(self.ngMesh.GetRegionNames(dim=2))
+                nBnd = len(self.ngMesh.GetRegionNames(dim=1))
                 if nMat > 1:
                     for e in self.ngMesh.Elements2D():
                         join = plex.getFullJoin([vStart+v.nr-1 for v in e.vertices])
                         plex.setLabelValue(CELL_SETS_LABEL, join[0], int(e.index))
                         cone  = plex.getCone(join[0])
                         for i in range(len(cone)):
-                            plex.setLabelValue(FACE_SETS_LABEL, cone[i], int(e.index))
+                            plex.setLabelValue(FACE_SETS_LABEL, cone[i], nBnd+int(e.index))
                 for e in self.ngMesh.Elements1D():
                     join = plex.getJoin([vStart+v.nr-1 for v in e.vertices])
-                    plex.setLabelValue(FACE_SETS_LABEL, join[0], nMat+int(e.index))
+                    plex.setLabelValue(FACE_SETS_LABEL, join[0], int(e.index))
                 self.petscPlex = plex
             else:
                 plex = PETSc.DMPlex().createFromCellList(2,
