### 2024-09-06 6655efa
Vanilla embedded Trefftz
Signed-off-by: Umberto Zerbinati <zerbinati@maths.ox.ac.uk>
