### 2024-09-06 a19653c
Added test for Trefftz methods
Signed-off-by: Umberto Zerbinati <zerbinati@maths.ox.ac.uk>
