### 2024-09-06 5ab079c
Finished implementing vanilla embedded Trefftz for polytopal method on mesh aggregation
Signed-off-by: Umberto Zerbinati <zerbinati@maths.ox.ac.uk>
