### 2024-09-07 27ab155
Using aggregation as a preconditioner
Signed-off-by: Umberto Zerbinati <zerbinati@maths.ox.ac.uk>
