+++ ngsPETSc/utils/firedrake/trefftz.py
@@ -7,15 +7,16 @@ except ImportError:
 from ngsPETSc.plex import CELL_SETS_LABEL, FACE_SETS_LABEL
 class TrefftzEmbedding(object):
 
-    def __init__(self, V, b, dim=None, tol=1e-12):
+    def __init__(self, V, b, dim=None, tol=1e-12, backend="scipy"):
         self.V = V
         self.b = b
         self.dim = V.dim() if not dim else dim + 1
         self.tol = tol
+        self.backend = backend
     
-    def assemble(self, backend="PETSc"):
+    def assemble(self):
         self.B = fd.assemble(self.b).M.handle
-        if backend == "scipy":
+        if self.backend == "scipy":
             import scipy.sparse as sp
             indptr, indices, data = self.B.getValuesCSR()
             Bsp = sp.csr_matrix((data, indices, indptr), shape=self.B.getSize())
@@ -23,11 +24,19 @@ class TrefftzEmbedding(object):
             QT = sp.csr_matrix(VT[0:sum(sig<self.tol), :])
             QTpsc = PETSc.Mat().createAIJ(size=QT.shape, csr=(QT.indptr, QT.indices, QT.data))
             self.dimT = QT.shape[0]
+            self.sig = sig
             return QTpsc, sig
         
-    def assembledEmbeddedMatrix(self, a, backend="PETSc"):
+    def assembledEmbeddedMatrix(self, a):
         self.A = fd.assemble(a).M.handle
-        self.QT, _ = self.assemble(backend)
+        self.QT, _ = self.assemble()
+        self.Q = PETSc.Mat().createTranspose(self.QT)
+        pscQTAQ = self.QT @ self.A @ self.Q
+        return pscQTAQ
+    
+    def assembledEmbeddedMatrixFree(self, a):
+        self.A = fd.assemble(a).M.handle
+        self.QT, _ = self.assemble()
         pythonQTAQ = self.embeddedMatrixWrap(self.QT, self.A)
         pscQTAQ = PETSc.Mat().create(comm=PETSc.COMM_WORLD)
         pscQTAQ.setSizes(self.dimT, self.dimT)
@@ -35,6 +44,18 @@ class TrefftzEmbedding(object):
         pscQTAQ.setPythonContext(pythonQTAQ)
         pscQTAQ.setUp()
         return pscQTAQ
+
+    def assembledEmbeddedPreconditioner(self, a):
+        self.A = fd.assemble(a).M.handle
+        self.QT, _ = self.assemble()
+        pythonQTAQ = self.embeddedPreconditioner(self, a)
+        pscQTAQ = PETSc.Mat().create(comm=PETSc.COMM_WORLD)
+        pscQTAQ.setSizes(self.dim, self.dim)
+        pscQTAQ.setType("python")
+        pscQTAQ.setPythonContext(pythonQTAQ)
+        pscQTAQ.setUp()
+        return pscQTAQ
+
     def assembledEmbeddedLoad(self, L):
         self.L = fd.assemble(L)
         with self.L.dat.vec as w:
@@ -46,6 +67,10 @@ class TrefftzEmbedding(object):
         with u.dat.vec as w:
             self.QT.multTranspose(y, w)
         return u
+    def embedVec(self, y):
+        w = self.QT.createVecRight()
+        self.QT.multTranspose(y, w)
+        return w
 
         
     class embeddedMatrixWrap(object):
@@ -65,6 +90,28 @@ class TrefftzEmbedding(object):
             self.QT.multTranspose(X, Z)
             self.A.mult(Z, W)
             self.QT.mult(W, Y)
+
+    class embeddedPreconditioner(object):
+        """
+        This class wraps a PETSc Preconditioner as PETSc Python matrix
[…]
