### 2024-09-07 e016a77
Aggregation+Treffetz
Signed-off-by: Umberto Zerbinati <zerbinati@maths.ox.ac.uk>
