+++ .github/workflows/ngsPETSc.yml
@@ -109,7 +109,7 @@ jobs:
       - name: Install Netgen and ngsPETSc
         run: |
           . /home/firedrake/firedrake/bin/activate
-          pip install netgen-mesher
+          pip install --pre netgen-mesher
           pip install xdist pytest-timeout ipympl
           NGSPETSC_NO_INSTALL_REQUIRED=ON pip install .
 
