### 2024-09-11 daca457
Post processing flags added
Signed-off-by: Umberto Zerbinati <zerbinati@maths.ox.ac.uk>
