### 2024-09-11 afb4d9d
Added tollerance parameter in post process curving
Signed-off-by: Umberto Zerbinati <zerbinati@maths.ox.ac.uk>
