### 2024-09-11 0d9de86
Removed post processing
Signed-off-by: Umberto Zerbinati <zerbinati@maths.ox.ac.uk>
