### 2024-09-12 5472275
Better way of constructing boundary array
Signed-off-by: Umberto Zerbinati <zerbinati@maths.ox.ac.uk>
