### 2024-09-12 1c2938f
Fix workflow for serial
Signed-off-by: Umberto Zerbinati <zerbinati@maths.ox.ac.uk>
