### 2024-09-12 1c00fe3
Fix for DG facet labels
Signed-off-by: Umberto Zerbinati <zerbinati@maths.ox.ac.uk>
