### 2024-09-12 a19dfd1
Fix for ngsolve 3D plex support
Signed-off-by: Umberto Zerbinati <zerbinati@maths.ox.ac.uk>
