### 2024-09-12 f9b2206
Making the linter happy
Signed-off-by: Umberto Zerbinati <zerbinati@maths.ox.ac.uk>
