+++ ngsPETSc/utils/firedrake/trefftz.py
@@ -27,14 +27,14 @@ class TrefftzEmbedding(object):
             self.sig = sig
             return QTpsc, sig
         
-    def assembledEmbeddedMatrix(self, a):
+    def embeddedMatrix(self, a):
         self.A = fd.assemble(a).M.handle
         self.QT, _ = self.assemble()
         self.Q = PETSc.Mat().createTranspose(self.QT)
         pscQTAQ = self.QT @ self.A @ self.Q
         return pscQTAQ
     
-    def assembledEmbeddedMatrixFree(self, a):
+    def embeddedMatrixAction(self, a):
         self.A = fd.assemble(a).M.handle
         self.QT, _ = self.assemble()
         pythonQTAQ = self.embeddedMatrixWrap(self.QT, self.A)
@@ -45,7 +45,7 @@ class TrefftzEmbedding(object):
         pscQTAQ.setUp()
         return pscQTAQ
 
-    def assembledEmbeddedPreconditioner(self, a):
+    def embeddedPreconditioner(self, a):
         self.A = fd.assemble(a).M.handle
         self.QT, _ = self.assemble()
         pythonQTAQ = self.embeddedPreconditioner(self, a)
@@ -56,7 +56,7 @@ class TrefftzEmbedding(object):
         pscQTAQ.setUp()
         return pscQTAQ
 
-    def assembledEmbeddedLoad(self, L):
+    def embeddedLoad(self, L):
         self.L = fd.assemble(L)
         with self.L.dat.vec as w:
             y =  self.QT.createVecLeft()
@@ -97,7 +97,7 @@ class TrefftzEmbedding(object):
         """
         def __init__(self, E, a):
             self.E = E
-            self.QTAQ = self.E.assembledEmbeddedMatrix(a)
+            self.QTAQ = self.E.embeddedMatrix(a)
             self.ksp = PETSc.KSP().create()
             self.ksp.setOperators(self.QTAQ)
             self.ksp.getPC().setType("lu")
@@ -114,19 +114,24 @@ class TrefftzEmbedding(object):
             self.E.embedVec(eY).copy(Y)
 
 class AggregationEmbedding(TrefftzEmbedding):
-    def __init__(self, V, mesh,b = None, dim=None, tol=1e-12):
+    def __init__(self, V, mesh, polyMesh, dim=None, tol=1e-12):
         # Relabel facets that are inside an aggregated region
+        offset = len(mesh.netgen_mesh.GetRegionNames(dim=1))\
+               + len(mesh.netgen_mesh.GetRegionNames(dim=2))
+        nPoly = int(max(polyMesh.dat.data[:])) # Number of aggregates
+        getIdx = mesh._cell_numbering.getOffset
         plex = mesh.topology_dm
         pStart,pEnd = plex.getDepthStratum(2)
-        numberBnd = len(mesh.netgen_mesh.GetRegionNames(dim=1))
-        numberMat = len(mesh.netgen_mesh.GetRegionNames(dim=2))
-        for mat in range(numberMat):
+        self.facet_index = []
+        for poly in range(nPoly+1):
             facets = []
             for i in range(pStart,pEnd):
-                if plex.getLabelValue(CELL_SETS_LABEL,i) == mat+1:
+                if polyMesh.dat.data[getIdx(i)] == poly:
                     for f in plex.getCone(i):
                         if f in facets:
-                            plex.setLabelValue(FACE_SETS_LABEL,f,numberBnd+numberMat+mat+1)
+                            plex.setLabelValue(FACE_SETS_LABEL,f,offset+poly)
+                            if offset+poly not in self.facet_index:
[…]
