### 2024-09-13 30096fa
WIP Polygonal of any order
Signed-off-by: Umberto Zerbinati <zerbinati@maths.ox.ac.uk>
