### 2024-09-13 209ab6c
Added test for Treftz aggregation and Treftz ksp
Signed-off-by: Umberto Zerbinati <zerbinati@maths.ox.ac.uk>
