+++ tests/test_trefftz.py
@@ -47,3 +47,40 @@ def test_trefftz_laplace():
     #assmeble the error
     assert(assemble(inner(uDG-uTF,uDG-uTF)*dx) < 1e-6)
     assert(E.dimT < V.dim()/2)
+
+def test_trefftz_aggregation():
+    from netgen.occ import WorkPlane, OCCGeometry
+    from ngsPETSc import dumbAggregation, AggregationEmbedding, trefftz_ksp
+
+    Rectangle = WorkPlane().Rectangle(1,1).Face()
+    geo = OCCGeometry(Rectangle, dim=2)
+    ngmesh = geo.GenerateMesh(maxh=0.3)
+    mesh = Mesh(ngmesh)
+
+    polymesh = dumbAggregation(mesh)
+
+    order = 3
+    x, y = SpatialCoordinate(mesh)
+    h = CellDiameter(mesh)
+    n = FacetNormal(mesh)
+    V = FunctionSpace(mesh, "DG", order)
+    u = TrialFunction(V)
+    v = TestFunction(V)
+    alpha = 1e3
+    mean_dudn = 0.5 * dot(grad(u("+"))+grad(u("-")),n("+"))
+    mean_dvdn = 0.5 * dot(grad(v("+"))+grad(v("-")),n("+"))
+    aDG = inner(grad(u),grad(v))* dx 
+    aDG += inner((alpha*order**2/(h("+")+h("-")))*jump(u),jump(v))*dS
+    aDG += inner(-mean_dudn,jump(v))*dS-inner(mean_dvdn,jump(u))*dS
+    aDG += alpha*order**2/h*inner(u,v)*ds
+    aDG += -inner(dot(n,grad(u)),v)*ds -inner(dot(n,grad(v)),u)*ds
+    f = Function(V).interpolate(exp(x)*sin(y)) 
+    L = alpha*order**2/h*inner(f,v)*ds - inner(dot(n,grad(v)),f)*ds
+
+    agg_embd = AggregationEmbedding(V, mesh, polymesh)
+    appctx = {"trefftz_embedding": agg_embd}
+
+    uDG = Function(V)
+    solve(aDG == L, uDG, solver_parameters={"ksp_type":"python","ksp_python_type":trefftz_ksp},appctx=appctx)
+
+    assert(assemble(inner(uDG-f,uDG-f)*dx)<1e-9)
