### 2024-09-13 01ad566
Using trefftz KSP also for Laplace test
Signed-off-by: Umberto Zerbinati <zerbinati@maths.ox.ac.uk>
