+++ ngsPETSc/utils/firedrake/trefftz.py
@@ -1,21 +1,34 @@
+"""
+This module provides a class to compute the Trefftz embedding of a given function space.
+It is also used to compute aggregation embedding of a given function space.
+"""
 try:
     import firedrake as fd
     from petsc4py import PETSc
-    from firedrake.__future__ import interpolate
 except ImportError:
     fd = None
-from ngsPETSc.plex import CELL_SETS_LABEL, FACE_SETS_LABEL
+from ngsPETSc.plex import FACE_SETS_LABEL
 
 class TrefftzEmbedding(object):
-
+    """
+    This class computes the Trefftz embedding of a given function space
+    :arg V: the function space
+    :arg b: the bilinear form defining the embedding
+    :arg dim: the dimension of the embedding
+    :arg tol: the tolerance for the singular values
+    :arg backend: the backend to use for the computation of the SVD
+    """
     def __init__(self, V, b, dim=None, tol=1e-12, backend="scipy"):
         self.V = V
         self.b = b
         self.dim = V.dim() if not dim else dim + 1
         self.tol = tol
         self.backend = backend
-    
+
     def assemble(self):
+        """
+        Assemble the embedding, compute the SVD and return the embedding matrix
+        """
         self.B = fd.assemble(self.b).M.handle
         if self.backend == "scipy":
             import scipy.sparse as sp
@@ -26,16 +39,27 @@ class TrefftzEmbedding(object):
             QTpsc = PETSc.Mat().createAIJ(size=QT.shape, csr=(QT.indptr, QT.indices, QT.data))
             self.dimT = QT.shape[0]
             self.sig = sig
-            return QTpsc, sig
+        else:
+            raise NotImplementedError("Only scipy backend is supported")
+        return QTpsc, sig
 
     def embeddedMatrix(self, a):
+        """
+        Compute the Trefftz embedding of the bilinear form a
+        :arg a: the bilinear form
+        """
         self.A = fd.assemble(a).M.handle
         self.QT, _ = self.assemble()
         self.Q = PETSc.Mat().createTranspose(self.QT)
         pscQTAQ = self.QT @ self.A @ self.Q
         return pscQTAQ
-    
+
     def embeddedMatrixAction(self, a):
+        """
+        Compute the Trefftz embedding of the bilinear form a,
+        and wrap it as a PETSc Python matrix.
+        :arg a: the bilinear form
+        """
         self.A = fd.assemble(a).M.handle
         self.QT, _ = self.assemble()
         pythonQTAQ = self.embeddedMatrixWrap(self.QT, self.A)
@@ -46,33 +70,34 @@ class TrefftzEmbedding(object):
         pscQTAQ.setUp()
         return pscQTAQ
 
-    def embeddedPreconditioner(self, a):
-        self.A = fd.assemble(a).M.handle
-        self.QT, _ = self.assemble()
-        pythonQTAQ = self.embeddedPreconditioner(self, a)
-        pscQTAQ = PETSc.Mat().create(comm=PETSc.COMM_WORLD)
-        pscQTAQ.setSizes(self.dim, self.dim)
-        pscQTAQ.setType("python")
-        pscQTAQ.setPythonContext(pythonQTAQ)
-        pscQTAQ.setUp()
-        return pscQTAQ
-
     def embeddedLoad(self, L):
+        """
+        Compute the Trefftz embedding of the load vector L
+        :arg L: the load vector
+        """
         self.L = fd.assemble(L)
         with self.L.dat.vec as w:
             y =  self.QT.createVecLeft()
             self.QT.mult(w, y)
         return y
     def embed(self, y):
[…]
