+++ .github/workflows/ngsPETSc.yml
@@ -120,3 +120,9 @@ jobs:
           pytest tests/regression/test_poisson_strong_bcs.py
           pytest tests/multigrid/test_netgen_gmg.py
           pytest tests/regression/test_netgen.py
+      
+      - name: Run extra ngsPETSc tests
+        run: |
+          . /home/firedrake/firedrake/bin/activate
+          cd $VIRTUAL_ENV/src/ngsPETSc
+          pytest tests/test_trefftz.py
