### 2024-09-19 1d924eb
Fix snapping to coarse thanks to Pablo
Signed-off-by: Umberto Zerbinati <zerbinati@maths.ox.ac.uk>
