+++ ngsPETSc/utils/firedrake/hierarchies.py
@@ -5,6 +5,8 @@ try:
     import firedrake as fd
     from firedrake.cython import mgimpl as impl
     from firedrake.__future__ import interpolate
+    import firedrake.dmhooks as dmhooks
+    import ufl
 except ImportError:
     fd = None
 
@@ -36,16 +38,79 @@ def snapToCoarse(coarse, linear, degree):
     '''
     This function snaps the coordinates of a DMPlex mesh to the coordinates of a Netgen mesh.
     '''
-    print(linear.coordinates.dat.data.shape)
-    if coarse.geometric_dimension() == 2:
-        low_order_element = linear.coordinates.function_space().ufl_element().sub_elements[0]
-        element = low_order_element.reconstruct(degree=degree)
-        space = fd.VectorFunctionSpace(linear, fd.BrokenElement(element))
+    dim = coarse.geometric_dimension()
+    if dim == 2:
+        coarseSpace = coarse.coordinates.function_space()
+        space = fd.VectorFunctionSpace(linear, "CG", degree)
+        bnd =space.boundary_nodes("on_boundary")
         ho = fd.assemble(interpolate(linear.coordinates, space))
-        bnd = coarse.coordinates.function_space().boundary_nodes("on_boundary")
-        print(bnd)
-        for i, pt in enumerate(ho.dat.data):
-            pass
+        coarseBnd = coarseSpace.boundary_nodes("on_boundary")
+        for i in bnd:
+            pt = ho.dat.data[i]
+            j = np.argmin(np.sum((coarse.coordinates.dat.data[coarseBnd] - pt)**2, axis=1))
+            ho.dat.data[i] = coarse.coordinates.dat.data[coarseBnd][j]
+        #Hyperelastic Smoothing
+        bcs = [fd.DirichletBC(space, ho, "on_boundary")]
+        quad_degree = 2*(degree+1)-1
+        dx = fd.dx(degree=quad_degree, domain=linear)
+        d = linear.topological_dimension()
+
+        Q = fd.TensorFunctionSpace(linear, "DG", degree=0)
+        Jinv = ufl.JacobianInverse(linear)
+        hinv = fd.Function(Q)
+        hinv.interpolate(Jinv)
+        G = ufl.Jacobian(linear) * hinv
+        ijac = 1/abs(ufl.det(G))
+        ref_grad = lambda u: ufl.dot(ufl.grad(u), G)
+        params = {
+            "snes_type": "newtonls",
+            "snes_linesearch_type": "l2",
+            "snes_max_it": 50,
+            "snes_rtol": 1E-8,
+            "snes_atol": 1E-8,
+            "snes_ksp_ew": True,
+            "snes_ksp_ew_rtol0": 1E-2,
+            "snes_ksp_ew_rtol_max": 1E-2,
+        }
+        params["mat_type"] = "aij"
+        coarse = {
+            "ksp_type": "preonly",
+            "pc_type": "lu",
+            "pc_mat_factor_type": "mumps",
+        }
+        gmg = {
+            "pc_type": "mg",
+            "mg_coarse": coarse,
+            "mg_levels": {
+                "ksp_max_it": 2,
+                "ksp_type": "chebyshev",
+                "pc_type": "jacobi",
+            },
+        }
+        l = fd.mg.utils.get_level(linear)[1]
+        pc = gmg if l else coarse
+        params.update(pc)
+        ksp = {
+            "ksp_rtol": 1E-8,
+            "ksp_atol": 0,
+            "ksp_type": "minres",
+            "ksp_norm_type": "preconditioned",
+        }
+        params.update(ksp)
+        u = ho
+        F = ref_grad(u)
+        J = ufl.det(F)
[…]
