### 2024-09-22 3a47b3a
Fix Docker container, to use a more recent image of ubuntu
Signed-off-by: Umberto Zerbinati <zerbinati@maths.ox.ac.uk>
