### 2024-09-22 3ae0065
Interpolating ratehr than wired snapping. Much better !!! Thanks @pef
Signed-off-by: Umberto Zerbinati <zerbinati@maths.ox.ac.uk>
