### 2024-09-22 e555ed6
Monthly Maintenance (#45)
* Fix Docker container, to use a more recent image of ubuntu

Signed-off-by: Umberto Zerbinati <zerbinati@maths.ox.ac.uk>

* Making the linter happy again, by ingoring error R0917:
Too many positional arguments

Signed-off-by: Umberto Zerbinati <zerbinati@maths.ox.ac.uk>

---------

Signed-off-by: Umberto Zerbinati <zerbinati@maths.ox.ac.uk>
Co-authored-by: Umberto Zerbinati <zerbinati@maths.ox.ac.uk>
