### 2024-09-23 1272911
Firedrake only has object of codim 1. I'm an idiot
Signed-off-by: Umberto Zerbinati <zerbinati@maths.ox.ac.uk>
