### 2024-09-25 e731c82
Nestededness of the snapping back to the geometry
Signed-off-by: Umberto Zerbinati <zerbinati@maths.ox.ac.uk>
