+++ ngsPETSc/utils/firedrake/meshes.py
@@ -78,51 +78,57 @@ def curveField(self, order, tol=1e-8):
 
     '''
     #Checking if the mesh is a surface mesh or two dimensional mesh
-    surf = len(self.netgen_mesh.Elements3D()) == 0
+    surf = len(self.netgen_mesh.Elements3D()) == 0 # REMOVE
+    if len(self.netgen_mesh.Elements3D()) == 0:
+        ng_element = self.netgen_mesh.Elements2D
+    else:
+        ng_element = self.netgen_mesh.Elements3D
+    ng_dimension = len(ng_element())
+
     #Constructing mesh as a function
     low_order_element = self.coordinates.function_space().ufl_element().sub_elements[0]
-    element = low_order_element.reconstruct(degree=order)
-    space = fd.VectorFunctionSpace(self, fd.BrokenElement(element))
-    newFunctionCoordinates = fd.assemble(interpolate(self.coordinates, space))
+    ufl_element = low_order_element.reconstruct(degree=order)
+    firedrake_space = fd.VectorFunctionSpace(self, fd.BrokenElement(ufl_element))
+    newFunctionCoordinates = fd.assemble(interpolate(self.coordinates, firedrake_space))
+
     #Computing reference points using fiat
     fiat_element = newFunctionCoordinates.function_space().finat_element.fiat_equivalent
     entity_ids = fiat_element.entity_dofs()
     nodes = fiat_element.dual_basis()
-    refPts = []
+    ref = []
     for dim in entity_ids:
         for entity in entity_ids[dim]:
             for dof in entity_ids[dim][entity]:
                 # Assert singleton point for each node.
                 pt, = nodes[dof].get_point_dict().keys()
-                refPts.append(pt)
-    V = newFunctionCoordinates.dat.data
-    refPts = np.array(refPts)
-    els = {True: self.netgen_mesh.Elements2D, False: self.netgen_mesh.Elements3D}
+                ref.append(pt)
+    reference_space_points = np.array(ref)
+
     #Mapping to the physical domain
+    els = {True: self.netgen_mesh.Elements2D, False: self.netgen_mesh.Elements3D} # REMOVE
+    physical_space_points = np.ndarray((ng_dimension, reference_space_points.shape[0], self.geometric_dimension()))
+    curved_space_points = np.ndarray((ng_dimension, reference_space_points.shape[0], self.geometric_dimension()))
+
     if self.comm.rank == 0:
-        physPts = np.ndarray((len(els[surf]()),
-                                refPts.shape[0], self.geometric_dimension()))
-        self.netgen_mesh.CalcElementMapping(refPts, physPts)
-        #Cruving the mesh
+        #Curving the mesh on rank 0
+        self.netgen_mesh.CalcElementMapping(reference_space_points, physical_space_points)
         self.netgen_mesh.Curve(order)
-        curvedPhysPts = np.ndarray((len(els[surf]()),
-                                    refPts.shape[0], self.geometric_dimension()))
-        self.netgen_mesh.CalcElementMapping(refPts, curvedPhysPts)
-        curved = els[surf]().NumPy()["curved"]
+        self.netgen_mesh.CalcElementMapping(reference_space_points, curved_space_points)
+        curved = ng_element().NumPy()["curved"]
     else:
-        physPts = np.ndarray((len(els[surf]()),
-                                refPts.shape[0], self.geometric_dimension()))
-        curvedPhysPts = np.ndarray((len(els[surf]()),
-                                    refPts.shape[0], self.geometric_dimension()))
-        curved = np.array((len(els[surf]()),1))
-    physPts = self.comm.bcast(physPts, root=0)
-    curvedPhysPts = self.comm.bcast(curvedPhysPts, root=0)
+        curved = np.array((ng_dimension, 1))
+
+    # Broadcast curving
[…]
