### 2024-09-25 3d12d14
Monthly maintenance (#46)
Lighter docker container
---------

Signed-off-by: Umberto Zerbinati <zerbinati@maths.ox.ac.uk>
Co-authored-by: Umberto Zerbinati <zerbinati@maths.ox.ac.uk>
