### 2024-09-26 814ff66
WIP: almost loop free
