### 2024-10-01 4bb63af
Additional tidying, difficult numpy indexing rules
