+++ ngsPETSc/utils/firedrake/meshes.py
@@ -12,6 +12,12 @@ except ImportError:
 import numpy as np
 from petsc4py import PETSc
 
+try:
+    from scipy.spatial.distance import cdist
+    HAVE_SCIPY = True
+except ImportError:
+    HAVE_SCIPY = False
+
 import netgen
 import netgen.meshing as ngm
 from netgen.meshing import MeshingParameters
@@ -70,6 +76,45 @@ def refineMarkedElements(self, mark):
     else:
         raise NotImplementedError("No implementation for dimension other than 2 and 3.")
 
+
+def _slow_cdist(XA, XB):
+        dist = np.zeros([len(XA), len(XB)])
+        for ii, a in enumerate(XA):
+            for jj, b in enumerate(XB):
+                dist[ii, jj] = np.linalg.norm(b - a)
+        return dist
+
+
+if not HAVE_SCIPY:
+    cdist = _slow_cdist
+
+
+def find_permutation(points_a, points_b, tol=1e-5):
+    """ Find all permutations between a list of two sets of points.
+
+    Given two numpy arrays of shape (ncells, npoints, dim) containing
+    floating point coordinates for each cell, determine each index
+    permutation that takes `points_a` to `points_b`. Ie:
+    ```
+    permutation = find_permutation(points_a, points_b)
+    assert np.allclose(points_a[permutation], points_b, rtol=0, atol=tol)
+    ```
+    """
+    if points_a.shape != points_b.shape:
+        raise ValueError("`points_a` and `points_b` must have the same shape.")
+
+    p = [np.where(cdist(a, b).T < tol)[1] for a, b in zip(points_a, points_b)]
+    try:
+        permutation = np.array(p, ndmin=2)
+    except ValueError:
+        raise ValueError("It was not possible to find a permutation for every cell within the provided tolerance")
+
+    if permutation.shape != points_a.shape[0:2]:
+        raise ValueError("It was not possible to find a permutation for every cell within the provided tolerance")
+
+    return permutation
+
+
 def curveField(self, order, tol=1e-8):
     '''
     This method returns a curved mesh as a Firedrake function.
@@ -78,21 +123,21 @@ def curveField(self, order, tol=1e-8):
 
     '''
     #Checking if the mesh is a surface mesh or two dimensional mesh
-    surf = len(self.netgen_mesh.Elements3D()) == 0 # REMOVE
     if len(self.netgen_mesh.Elements3D()) == 0:
         ng_element = self.netgen_mesh.Elements2D
     else:
         ng_element = self.netgen_mesh.Elements3D
     ng_dimension = len(ng_element())
+    geom_dim = self.geometric_dimension()
 
     #Constructing mesh as a function
     low_order_element = self.coordinates.function_space().ufl_element().sub_elements[0]
     ufl_element = low_order_element.reconstruct(degree=order)
     firedrake_space = fd.VectorFunctionSpace(self, fd.BrokenElement(ufl_element))
-    newFunctionCoordinates = fd.assemble(interpolate(self.coordinates, firedrake_space))
+    new_coordinates = fd.assemble(interpolate(self.coordinates, firedrake_space))
 
     #Computing reference points using fiat
-    fiat_element = newFunctionCoordinates.function_space().finat_element.fiat_equivalent
+    fiat_element = new_coordinates.function_space().finat_element.fiat_equivalent
     entity_ids = fiat_element.entity_dofs()
     nodes = fiat_element.dual_basis()
     ref = []
@@ -105,9 +150,8 @@ def curveField(self, order, tol=1e-8):
     reference_space_points = np.array(ref)
 
     #Mapping to the physical domain
-    els = {True: self.netgen_mesh.Elements2D, False: self.netgen_mesh.Elements3D} # REMOVE
[…]
