### 2024-10-01 f3ad982
More notes
