### 2024-10-03 9e1e9cd
Tolerance needs to be dropped for cell location
