### 2024-10-03 e45542c
Remove even more parallel communication
