### 2024-10-03 7b4f590
Add PETSc event decorators
