### 2024-10-04 3831f7c
Prefer np.zeros over np.ndarray
