### 2024-10-04 d986ee0
Different tollerance for permutation and barycenter location
