### 2024-10-04 0ecbfe7
Scipy dependencies in setup.py
