### 2024-10-04 fe5a81e
Only build docker on main branch (#49)
Thanks @JDBetteridge
