### 2024-10-04 14f2719
Move stuff to pyproject.toml
