+++ setup.py
@@ -1,10 +1,5 @@
 import setuptools, os
 
-with open("README.md", "r", encoding = "utf-8") as fh:
-    long_description = fh.read()
-    
-scipy = '' if 'NGSPETSC_NO_SCIPY' in os.environ else 'scipy'
-    
 if 'NGSPETSC_NO_INSTALL_REQUIRED' in os.environ:
     install_requires = []
 elif 'NGS_FROM_SOURCE' in os.environ:
@@ -12,7 +7,7 @@ elif 'NGS_FROM_SOURCE' in os.environ:
         'petsc4py',
         'mpi4py',
         'numpy',
-        scipy,
+        'scipy',
         'pytest', #For testing
         'pylint', #For formatting
         ]
@@ -22,29 +17,12 @@ else:
         'ngsolve',
         'petsc4py',
         'mpi4py',
-        scipy,
+        'scipy',
         'pytest', #For testing
         'pylint', #For formatting
     ]
 
 setuptools.setup(
-    name = "ngsPETSc",
-    version = "0.0.5",
-    author = "Umberto Zerbinati",
-    author_email = "umberto.zerbinati@maths.ox.ac.uk",
-    description = "NGSolve/Netgen interface to PETSc.",
-    long_description = long_description,
-    long_description_content_type = "text/markdown",
-    url = "",
-    project_urls = {
-    },
-    classifiers = [
-        "Programming Language :: Python :: 3",
-        "License :: OSI Approved :: MIT License",
-        "Operating System :: OS Independent",
-    ],
-    packages=["ngsPETSc", "ngsPETSc.utils", "ngsPETSc.utils.firedrake", "ngsPETSc.utils.ngs"],
-    python_requires = ">=3.8",
-    install_requires=install_requires
-
+    install_requires=install_requires,
+    packages=["ngsPETSc", "ngsPETSc.utils", "ngsPETSc.utils.firedrake", "ngsPETSc.utils.ngs"]
 )
