### 2024-10-04 4874b86
Remove slow cdist as scipy is now a hard dependency
