### 2024-10-08 dedddf7
JDBetteridge/refactor curvefield (#48)
Parallel safe curvefield. Thank you @JDBetteridge !
