+++ ngsPETSc/utils/firedrake/meshes.py
@@ -1,6 +1,6 @@
 '''
 This module contains all the functions related to wrapping NGSolve meshes to Firedrake
-We adopt the same docstring conventiona as the Firedrake project, since this part of
+We adopt the same docstring conventions as the Firedrake project, since this part of
 the package will only be used in combination with Firedrake.
 '''
 try:
@@ -11,6 +11,7 @@ except ImportError:
 
 import numpy as np
 from petsc4py import PETSc
+from scipy.spatial.distance import cdist
 
 import netgen
 import netgen.meshing as ngm
@@ -70,105 +71,148 @@ def refineMarkedElements(self, mark):
     else:
         raise NotImplementedError("No implementation for dimension other than 2 and 3.")
 
-def curveField(self, order, tol=1e-8, CG=False):
+
+@PETSc.Log.EventDecorator()
+def find_permutation(points_a, points_b, tol=1e-5):
+    """ Find all permutations between a list of two sets of points.
+
+    Given two numpy arrays of shape (ncells, npoints, dim) containing
+    floating point coordinates for each cell, determine each index
+    permutation that takes `points_a` to `points_b`. Ie:
+    ```
+    permutation = find_permutation(points_a, points_b)
+    assert np.allclose(points_a[permutation], points_b, rtol=0, atol=tol)
+    ```
+    """
+    if points_a.shape != points_b.shape:
+        raise ValueError("`points_a` and `points_b` must have the same shape.")
+
+    p = [np.where(cdist(a, b).T < tol)[1] for a, b in zip(points_a, points_b)]
+    try:
+        permutation = np.array(p, ndmin=2)
+    except ValueError as e:
+        raise ValueError(
+            "It was not possible to find a permutation for every cell"
+            " within the provided tolerance"
+        ) from e
+
+    if permutation.shape != points_a.shape[0:2]:
+        raise ValueError(
+            "It was not possible to find a permutation for every cell"
+            " within the provided tolerance"
+        )
+
+    return permutation
+
+
+@PETSc.Log.EventDecorator()
+def curveField(self, order, permutation_tol=1e-8, location_tol=1e-1, cg_field=False):
     '''
     This method returns a curved mesh as a Firedrake function.
 
-    :arg order: the order of the curved mesh
+    :arg order: the order of the curved mesh.
+    :arg permutation_tol: tolerance used to construct the permutation of the reference element.
+    :arg location_tol: tolerance used to locate the cell a point belongs to.
+    :arg cg_field: return a CG function field representing the mesh, rather than the
+                   default DG field.
 
     '''
-    #Checking if the mesh is a surface mesh or two dimensional mesh
-    surf = len(self.netgen_mesh.Elements3D()) == 0
-    #Constructing mesh as a function
-    if CG:
-        space = fd.VectorFunctionSpace(self, "CG", order)
+    # Check if the mesh is a surface mesh or two dimensional mesh
+    if len(self.netgen_mesh.Elements3D()) == 0:
+        ng_element = self.netgen_mesh.Elements2D
+    else:
+        ng_element = self.netgen_mesh.Elements3D
+    ng_dimension = len(ng_element())
+    geom_dim = self.geometric_dimension()
+
+    # Construct the mesh as a Firedrake function
+    if cg_field:
+        firedrake_space = fd.VectorFunctionSpace(self, "CG", order)
     else:
         low_order_element = self.coordinates.function_space().ufl_element().sub_elements[0]
-        element = low_order_element.reconstruct(degree=order)
-        space = fd.VectorFunctionSpace(self, fd.BrokenElement(element))
-    newFunctionCoordinates = fd.assemble(interpolate(self.coordinates, space))
[…]
