### 2024-10-14 #7014 @danielskatz
👋 @schruste - Would you be willing to review this JOSS submission?
