### 2024-10-14 #7014 @rckirby
What are your conflict of interest guidelines?  I am a co-author/collaborator with one of the authors.
