### 2024-10-14 #7014 @danielskatz
@rckirby - see https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy  Basically, it depends on how long ago the collaboration ended (the work was done), assuming it is not current.
