### 2024-10-15 #7014 @danielskatz
👋 @thelfer & @mscroggs, would one or both of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
