### 2024-10-15 #7014 @rckirby
I've read the CoI policy and will decline.
