### 2024-10-15 #7359 @editorialbot
[REVIEW]: ngsPETSc: A coupling between NETGEN/NGSolve 2  and PETSc
**Submitting author:** @uzerbinati (<a href="http://orcid.org/0000-0002-2577-1106">Umberto Zerbinati</a>)
**Repository:** https://github.com/NGSolve/ngsPETSc
**Branch with paper.md** (empty if default branch):
**Version:** 0.0.8
