### 2024-10-16 #7359 @knepley
@danielskatz Looking at the repo, based on the number of commits, number of lines, and recently opened MRs, the authors should probably also include JDBetteridge in the author list. How do you normally approach this at JOSS?
