### 2024-10-16 #7359 @danielskatz
@knepley - This is really a question for the author, rather than for JOSS.

👋 @UZerbinati - can you say something about this?
