### 2024-10-16 #7359 @danielskatz
Yes! JOSS reviewers are meant to be interactive between the author(s) and reviewers, with my role being to keep things on track and progressing. Think of this like any other open source software discussion.
