### 2024-10-16 #7359 @knepley
@UZerbinati There is not really a State of the Field in the paper, which would be a brief list of packages with similar capabilities, maybe DUNE or FreeFEM?
