### 2024-10-16 #7359 @knepley
@UZerbinati I don't see guidelines for contributing. Perhaps it would be enough to point to the Firedrake guidelines in the documentation?
