### 2024-10-16 #7359 @UZerbinati
Dear  @danielskatz  and @knepley, the reason why Jack is not among the authors is the fact that most of his contributions were made after the submission of this manuscript. After talking with the other authors and Jack, we added Jack among the paper's authors.
