### 2024-10-16 f1fa686
Update install.rst to include Contribution guidelines
As suggested by @knepley in openjournals/joss-reviews#7359 we added contribution guidelines.
