### 2024-10-16 a196ed4
Update install.rst to include Contribution guidelines (#52)
As suggested by @knepley in openjournals/joss-reviews#7359 we added contribution guidelines.
