+++ paper/paper.bib
@@ -181,3 +181,59 @@ doi={10.1137/0907058}
  R.},
  year={2006},
  pages={245--269} }
+@article{FreeFEM,
+  title={FreeFem++ manual},
+  author={Hecht, Fr{\'e}d{\'e}ric and Pironneau, Olivier and Le Hyaric, A and Ohtsuka, K},
+  journal={Laboratoire Jacques Louis Lions},
+  year={2005}
+}
+@Article{DUNE,
+author={Dedner, Andreas
+and Kl{\"o}fkorn, Robert
+and Nolte, Martin
+and Ohlberger, Mario},
+title={A generic interface for parallel and adaptive discretization schemes: abstraction principles and the Dune-Fem module},
+journal={Computing},
+year={2010},
+month={Nov},
+day={01},
+volume={90},
+number={3},
+pages={165-196},
+issn={1436-5057},
+doi={10.1007/s00607-010-0110-3},
+url={https://doi.org/10.1007/s00607-010-0110-3}
+}
+@Article{dealII,
+  title   = {The \texttt{deal.II} Library, Version 9.5},
+  author  = {Daniel Arndt and Wolfgang Bangerth and Maximilian Bergbauer and
+             Marco Feder and Marc Fehling and Johannes Heinz and
+             Timo Heister and Luca Heltai and Martin Kronbichler and
+             Matthias Maier and Peter Munch and Jean-Paul Pelteret and
+             Bruno Turcksin and David Wells and Stefano Zampini},
+  journal = {Journal of Numerical Mathematics},
+  year    = {2023},
+  doi     = {10.1515/jnma-2023-0089},
+  pages   = {231--246},
+  volume  = {31},
+  number  = {3},
+  url     = {https://dealii.org/deal95-preprint.pdf}
+}
+@misc{dolfinX,
+  author       = {Baratta, Igor A. and
+                  Dean, Joseph P. and
+                  Dokken, Jørgen S. and
+                  Habera, Michal and
+                  Hale, Jack S. and
+                  Richardson, Chris N. and
+                  Rognes, Marie E. and
+                  Scroggs, Matthew W. and
+                  Sime, Nathan and
+                  Wells, Garth N.},
+  title        = {{DOLFINx: The next generation FEniCS problem solving environment}},
+  month        = dec,
+  year         = 2023,
+  publisher    = {Zenodo},
+  doi          = {10.5281/zenodo.10447666},
+  url          = {https://doi.org/10.5281/zenodo.10447666}
+}
