### 2024-11-01 NGSolve/ngsPETSc#54 @thelfer
[JOSS review] Installation link in `README.md` is broken
`https://ngspetsc.readthedocs.io/en/latest/installation.html` shall be replaced by `https://ngspetsc.readthedocs.io/en/latest/install.html`.
