### 2024-11-01 NGSolve/ngsPETSc#55 @thelfer
[JOSS Review] Feedback on the installation proccess
1. There is an extra `\` at the end of this command:
    ~~~
    cd petsc
    python configure --download-chaco \
   ...
    --with-petsc4py=1
    ~~~
2. Without installing exta dependencies, one need to pass `-DUSE_OCC=OFF` to `cmake` when building `ngsolve`
