### 2024-11-01 NGSolve/ngsPETSc#56 @thelfer
[JOSS REVIEW] first example fails
The following example, taken from https://ngspetsc.readthedocs.io/en/latest/PETScKSP/poisson.py.html, fails as follows:

~~~
$ python3 poisson.py
optfile ./ng.opt does not exist - using default values
togl-version : 2
loading ngsolve library
NGSolve-6.2.2405-28-g63cfc18ca
Using Lapack
Including sparse direct solver UMFPACK
Running parallel using 32 thread(s)
Traceback (most recent call last):
  File "/home/th202608/codes/ngsPETSc/tests/Poisson/poisson.py", line 7, in <module>
    ngmesh = unit_square.GenerateMesh(maxh=0.1)
NameError: name 'unit_square' is not defined
~~~
