### 2024-11-06 NGSolve/ngsPETSc#56 @UZerbinati
This example is running for me on a fresh install and on the docker image.
Could it be that something went wrong with your ngsolve install?
