### 2024-11-06 00695dd
Remove setuptools
