### 2024-11-06 NGSolve/ngsPETSc#56 @thelfer
I just followed your tutorial steps by steps. `ngsolve` did get installed, so I expect that there is a dependency missing.
