### 2024-11-06 NGSolve/ngsPETSc#56 @UZerbinati
Did you install it with `-DUSE_OCC=OFF` instead of the default `-DUSE_OCC=ON` ?
