### 2024-11-06 NGSolve/ngsPETSc#56 @thelfer
must check
