### 2024-11-06 NGSolve/ngsPETSc#56 @UZerbinati
Because `unit_square` has lately become an OCC object in solve hence, if you install without OCC support it might not be there.
You should be able to test the code even without OCC if you use `from netgen.geom2d import unit_square`.
