### 2024-11-06 NGSolve/ngsPETSc#56 @thelfer
Indeed. I remember that I removed it because OCC was **not** listed as a dependency to `ngsolve`
