### 2024-11-06 NGSolve/ngsPETSc#56 @thelfer
I think that you have two choices :)

- update the installation guide to include `OCC` as a mandatory dependency
- update your examples
