### 2024-11-06 f880395
Fix  Issue #56 (#59)
Fix Issue #56 

---------

Signed-off-by: Umberto Zerbinati <zerbinati@maths.ox.ac.uk>
Co-authored-by: Umberto Zerbinati <zerbinati@maths.ox.ac.uk>
