### 2024-11-06 NGSolve/ngsPETSc#56 @UZerbinati
We have included include OCC as a mandatory dependency (we will also change the pyproject.toml in next PR).
