### 2024-11-06 NGSolve/ngsPETSc#55 @UZerbinati
This should have been fixed.
