### 2024-11-06 NGSolve/ngsPETSc#54 @UZerbinati
This should have been fixed.
