### 2024-11-06 7d1b9bc
Fix typing annotation of output type of GeometricModel in utils/fenicsx.py (#60)
Co-authored-by: Francesco Ballarin <francesco.ballarin@unicatt.it>
