### 2024-11-06 615e26c
Add info from poetry, but not petsc4py
