### 2024-11-06 8689875
Add missing petsc4py dependency
