### 2024-11-06 fe4ff9e
Remove duplicated metadata
