### 2024-11-06 3f1d91c
Remove ineffective environment variables
