### 2024-11-07 92b5cfb
Update paper.md
Updated text with Patrick suggestions
