### 2024-11-07 23673c0
Update paper.bib
Updated bibliography
