+++ paper/paper.bib
@@ -24,6 +24,23 @@
   year={2023},
   doi= {10.25561/104839}
 }
+@article{TETGEN,
+author = {Si, Hang},
+title = {TetGen, a Delaunay-Based Quality Tetrahedral Mesh Generator},
+year = {2015},
+issue_date = {January 2015},
+publisher = {Association for Computing Machinery},
+address = {New York, NY, USA},
+volume = {41},
+number = {2},
+issn = {0098-3500},
+url = {https://doi.org/10.1145/2629697},
+doi = {10.1145/2629697},
+journal = {ACM Trans. Math. Softw.},
+month = feb,
+articleno = {11},
+numpages = {36}
+}
 @techreport{PETSc,
   author={Satish Balay and Shrirang Abhyankar and Mark~F. Adams and Steven Benson and Jed
           Brown and Peter Brune and Kris Buschelman and Emil Constantinescu and Lisandro
@@ -236,4 +253,12 @@ url={https://doi.org/10.1007/s00607-010-0110-3}
   publisher    = {Zenodo},
   doi          = {10.5281/zenodo.10447666},
   url          = {https://doi.org/10.5281/zenodo.10447666}
+}
+@article{GMSH,
+  title={A three-dimensional finite element mesh generator with built-in pre-and post-processing facilities},
+  author={Geuzaine, Christophe and Remacle, Jean-Fran{\c{c}}ois},
+  journal={International Journal for Numerical Methods in Engineering},
+  volume={11},
+  pages={79},
+  year={2020}
 }
