### 2024-11-07 347825f
Update paper.md
Fix affiliation
