### 2024-11-07 dfaf336
Update paper.md
Fix affiliation a second time
