+++ pyproject.toml
@@ -12,7 +12,7 @@ firedrake = [
 
 [tool.poetry]
 name = "ngsPETSc"
-version = "0.0.6"
+version = "0.0.7a0"
 description = "NGSolve/Netgen interface to PETSc"
 authors = [
   "Umberto Zerbinati <umberto.zerbinati@maths.ox.ac.uk>",
@@ -24,7 +24,7 @@ maintainers = [
   "Umberto Zerbinati <umberto.zerbinati@maths.ox.ac.uk>",
 ]
 readme = "README.md"
-licence = "MIT"
+license = "MIT"
 packages = [
     {include = "ngsPETSc"}
 ]
