### 2024-11-07 49ba98f
Try building petsc/petsc4py
