+++ pyproject.toml
@@ -12,7 +12,7 @@ firedrake = [
 
 [tool.poetry]
 name = "ngsPETSc"
-version = "0.0.7a1"
+version = "0.0.7a2"
 description = "NGSolve/Netgen interface to PETSc"
 authors = [
   "Umberto Zerbinati <umberto.zerbinati@maths.ox.ac.uk>",
