### 2024-11-07 fba882d
PETSc requires additional Python packages to install petsc4py
