### 2024-11-07 a75c139
export PETSc env vars
