+++ .github/workflows/release.yml
@@ -36,7 +36,7 @@ jobs:
             echo "new_version=${NEW_VERSION:1}" >> "$GITHUB_OUTPUT"
             echo "suffix=$SUFFIX" >> "$GITHUB_OUTPUT"
             echo "tag_name=$TAG_NAME" >> "$GITHUB_OUTPUT"
-            echo "Version is $NEW_VERSION"
+            echo "Version is ${NEW_VERSION:1}"
             echo "Suffix is $SUFFIX"
             echo "Tag name is $TAG_NAME"
           else
