+++ pyproject.toml
@@ -1,15 +1,3 @@
-[project.optional-dependencies]
-dev = [
-    "pytest",
-    "pylint",
-]
-ngsolve = [
-    "ngsolve",
-]
-firedrake = [
-    "firedrake",
-]
-
 [tool.poetry]
 name = "ngsPETSc"
 version = "0.0.7a6"
@@ -43,7 +31,7 @@ netgen-mesher = "^6.2"
 netgen-occt = "^7.8"
 petsc4py =  "^3.22.1"
 mpi4py = "^4"
-numpy = "^2.0"
+numpy = "^2"
 scipy = "^1"
 
 [tool.poetry.group.dev.dependencies]
@@ -56,3 +44,8 @@ ngsolve = "^6.2"
 [build-system]
 requires = ["poetry-core"]
 build-backend = "poetry.core.masonry.api"
+
+[project.optional-dependencies]
+firedrake = [
+    "firedrake",
+]
