### 2024-11-08 4afb19e
A bit of SSH debugging
