+++ .github/workflows/release.yml
@@ -121,8 +121,13 @@ jobs:
         run: |
           export PETSC_DIR=/home/runner/work/ngsPETSc/petsc
           export PETSC_ARCH=arch-linux-c-opt
-          pip install ../petsc/src/binding/petsc4py
+          # Hopefully this song and dance goes away after petsc4py release 3.22.2
+          poetry remove petsc4py
           poetry install --sync --no-interaction
+          poetry add ../petsc/src/binding/petsc4py
+          sed -i -E "s/petsc4py = (\{.+\})/petsc4py = \"^3.22.1\"/" pyproject.toml
+          # Replace when fixed with:
+          # poetry install --sync --no-interaction
 
       - name: Build source and wheel distribution
         run: |
