### 2024-11-08 742be52
Remove SSH debugging
