### 2024-11-08 652d57b
Update the readme to prefer installing via PyPI
