### 2024-11-08 febeac3
Add notes on how to create a release for Umberto
