### 2024-11-12 #7359 @danielskatz
👋 @thelfer - How are things going now?
