### 2024-11-12 #7359 @thelfer
@danielskatz Slowly for sure.

I made issues on the installation process.

- https://github.com/NGSolve/ngsPETSc/issues/54
- https://github.com/NGSolve/ngsPETSc/issues/55
- https://github.com/NGSolve/ngsPETSc/issues/56

I must now recompile `ngsolve` with `OpenCascade` enabled for the tests to work.
