### 2024-11-14 89e8adc
Fixes for Cell.geometric_dimension
