### 2024-11-20 a645e06
Fixes for Cell.geometric_dimension (#63)
