+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "ngsPETSc"
-version = "0.0.7a7"
+version = "0.0.7a8"
 description = "NGSolve/Netgen interface to PETSc"
 authors = [
   "Umberto Zerbinati <umberto.zerbinati@maths.ox.ac.uk>",
