### 2024-11-22 #7359 @UZerbinati
@thelfer it might be helpful that we started distributing a wheel for ngsPETSc (it is still in alpha phase because wheels with petsc4py and mpi4py are always tricky): [PyPI](https://pypi.org/project/ngsPETSc/)
