### 2024-11-24 #7359 @thelfer
> it might be helpful that we started distributing a wheel for ngsPETSc

Thanks for the input. I'll have a look quickly.
