### 2024-11-27 NGSolve/ngsPETSc#56 @thelfer
Hi @UZerbinati,

Sorry for being so slow.

I reinstalled `ngsolve`, but there is still an issue with your install instructions: I had to add this package to be able to build `ngsolve` with `OpenCascade` enabled:

~~~
$ pip install  netgen-occt-devel
~~~

As a suggestion, I would add this line in your tutorial after the call to `Draw`:

~~~
input("Press the <ENTER> key to continue...")
~~~
