### 2024-12-02 98d1808
Monthly clean up
Signed-off-by: Umberto Zerbinati <zerbinati@maths.ox.ac.uk>
