### 2024-12-03 NGSolve/ngsPETSc#64 @thelfer
[JOSS] Review of the paper
Overall the paper is well written and easy to read. Except from a small remark below, the paper is fine for me.

However, the JOSS check list insists on two points *Data sharing* and *Reproducibility*. Are the tests presented in the paper available in the  `ngsPETSc` sources or in the `Firedrake` sources ?

# A small remark

The project has two seperate goals:

1. All PETSc solvers benefits from NETGEN meshes and geometries.
2. NGSolve users access to the wide array of linear, nonlinear solvers,
   and time-steppers available in PETSc.

The first statement shall be more precise as only solvers based on `DMPLEX` seem able to benefit from `ngsPETSc` (example: `Firedrake`). There are many solvers that use `PETSc` but don't use `DMPLEX` as their mesh backends (example: `FEniCS`). In fact, the paper only focus on `Firedrake` which somehow further weakens the genericity of this first
statement.

I think that the first statement shall be more humble without making it less interesting. I may propose something like:

> `ngsPETSc` seamlessly integrates with `PETSc` `DMPLEX`, a feature that may benefit to any solver based in this solution. This paper illustrates this in the case of the `Firedrake` solver.
