### 2024-12-03 #7359 @thelfer
@danielskatz I almost finished my review. There are still a few questions left regarding the availability of the tests cases exhibited in the paper: https://github.com/NGSolve/ngsPETSc/issues/64
