### 2024-12-03 e562edf
Fix to installation instructions, thanks to Thomas Helfer
