### 2024-12-03 87c208e
Edits to the paper suggested by Thomas Helfer
