### 2024-12-03 NGSolve/ngsPETSc#56 @UZerbinati
Thank you for spotting this. It has been fixed in PR #65.
