### 2024-12-03 NGSolve/ngsPETSc#64 @UZerbinati
Dear Thomas,

Thank you for taking the time to review our manuscript.

All examples in the paper are either in the ngsPETSc documentation `docs/src` or in `paper/examples`. In particular all code for reproducing the results in the manuscript is in the ngsPETSc repository.

We agree that we should modify our first statement and we have implemented your suggestion in PR #66 .
