### 2024-12-03 #7359 @thelfer
Issue https://github.com/NGSolve/ngsPETSc/issues/64 has been quickly fixed. On my side, everything is ok and the paper is ready to be published.
