### 2024-12-03 #7359 @danielskatz
👋 @thelfer

Note that the definition for Automated tests (which might be slightly misnamed) is "Are there automated tests or manual steps described so that the functionality of the software can be verified?"

And yes, the paper is required to discuss the state of the field - "Do the authors describe how this software compares to other commonly-used packages?"

If neither of these can be checked off by you currently, the authors will need to address them before we can say that your review is complete and the work can be published.
