### 2024-12-03 #7359 @thelfer
@danielskatz

>  "Are there automated tests or manual steps described so that the functionality of the software can be verified?"

Digging in the `Contributing` section of the documentation, there is a paragraph showing how to run the tests.
It works.

> "Do the authors describe how this software compares to other commonly-used packages?"

I would say that this point is questionnable.

The paper shows how useful it is for `NGSolve` and `Firedrake`, and what new features are brought to those solvers.

I don't even know if this criterion is applicable in this context.
