### 2024-12-03 #7359 @UZerbinati
@danielskatz @thelfer

Regarding automated testing, we also have a CI which was running smoothly until yesterday. I'm planning to find out the bug that is causing its failure this weekend. Here is the GitHub workflow for the CI: https://github.com/NGSolve/ngsPETSc/blob/main/.github/workflows/ngsPETSc.yml.

Regarding the state of the art, as suggested by @knepley we added a paragraph where we briefly list packages with similar capabilities, including DUNE and FreeFEM, here you can find the MR: https://github.com/NGSolve/ngsPETSc/pull/62
