### 2024-12-05 #7359 @knepley
> 👋 @knepley - (asking again) How are you doing on this? It looks like you just have a few items left to check off. I wonder if they have been addressed by the authors now.

Sorry, holidays are hectic. I have checked off the remaining items.
