### 2024-12-06 5fc5089
Edits to the paper suggested by Thomas Helfer (#66)
