+++ paper/paper.bib
@@ -262,3 +262,16 @@ url={https://doi.org/10.1007/s00607-010-0110-3}
   pages={79},
   year={2020}
 }
+
+@inproceedings{DMPLEX,
+author = {Lange, Michael and Knepley, Matthew G. and Gorman, Gerard J.},
+title = {Flexible, scalable mesh and data management using {PETSc DMPlex}},
+year = {2015},
+isbn = {9780992661519},
+publisher = {University of Edinburgh},
+address = {Great Britain},
+booktitle = {Proceedings of the 3rd International Conference on Exascale Applications and Software},
+pages = {71--76},
+location = {Edinburgh, UK},
+series = {EASC '15}
+}
