+++ paper/paper.md
@@ -33,7 +33,7 @@ authors:
     orcid: 0000-0002-2577-1106
     corresponding: true
     equal-contrib: true 
-    affiliation: 2
+    affiliation: 3
 affiliations:
  - name: Imperial College London, United Kingdom
    index: 1
