### 2024-12-06 5974ac9
adding missing heading to paper.md
