### 2024-12-06 7962553
minor changes in paper.bib
