+++ paper/paper.bib
@@ -3,7 +3,7 @@
   publisher = {Springer Science and Business Media LLC},
   volume = {1},
   year = {1997},
-  title = {NETGEN An advancing front 2D/3D-mesh generator based on abstract rules},
+  title = {NETGEN {A}n advancing front 2D/3D-mesh generator based on abstract rules},
   author = {Schöberl, Joachim},
   journal = {Computing and visualization in science},
   number = {1},
@@ -26,7 +26,7 @@
 }
 @article{TETGEN,
 author = {Si, Hang},
-title = {TetGen, a Delaunay-Based Quality Tetrahedral Mesh Generator},
+title = {TetGen, a {D}elaunay-Based Quality Tetrahedral Mesh Generator},
 year = {2015},
 issue_date = {January 2015},
 publisher = {Association for Computing Machinery},
@@ -136,11 +136,12 @@ year = {1978}
 }
 @TechReport{ml,
   AUTHOR = {M. Sala and J.J. Hu and R.S. Tuminaro},
-  TITLE = "{ML}3.1 {S}moothed {A}ggregation {U}ser's {G}uide",
+  TITLE = "{ML}3.1 Smoothed Aggregation User's Guide",
   INSTITUTION =  "Sandia National Laboratories",
   NUMBER = "SAND2004-4821",
   ADDRESS = "Albuquerque, NM (USA)",
-  YEAR = "2004"
+  YEAR = "2004",
+  doi = {10.2172/974895}
 }
 @article{Boffi,
 author = {D. Boffi},
