### 2024-12-06 #7359 @danielskatz
@UZerbinati - I have suggested some small changes in https://github.com/NGSolve/ngsPETSc/pull/67 - please merge this.

Also, I note there is no Acknowledgements section in your paper. Should there be, mentioning any funding or anything else?
