### 2024-12-07 ec3d189
Removed MPI direct dependency from ngsPETSc. MPI should either come from NGSolve/Firedrake or FEniCSx.
Signed-off-by: Umberto Zerbinati <zerbinati@maths.ox.ac.uk>
