### 2024-12-07 f3252f6
Build system added to CI
