### 2024-12-09 #7359 @UZerbinati
Dear @danielskatz,
we have updated the paper to include an acknowledgements section. The DOI of the archived version is 10.5281/zenodo.14295034.
