### 2024-12-09 #7359 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
